Fix '_max_penetrance' with new reduction strategy
The expressions must be monotonically non-decreasing and sortable when
'x' and 'y' are real positive numbers. So, to achieve the largest
polynomial within the penetrance expressions, simply does a substitution
for 'x' and 'y' to real positive numbers and the largest numerical
reduction of the expressions will be also the largest expression.

This strategy is pending to be accepted by directors.