Disable prev. or her. config till model is loaded
Actually this parameter is independent of the load of the model.
However, thus limiting the order of filling fields, it is possible to
validate this field with the validator of the model, which otherwise
might not be available. It also makes the user better understand the
order in which the interface should be used.